+++ includes/classes/admin/class-setup.php
@@ -136,16 +136,18 @@ class Setup
 				'type'         => 'boolean',
 				'single'       => true,
 				'show_in_rest' => true,
+				'default'      => false,
 			)
 		);
 
 		register_meta(
 			'post',
-			'disable_single',
+			'lsx_to_hide_from_listings',
 			array(
 				'type'         => 'boolean',
 				'single'       => true,
 				'show_in_rest' => true,
+				'default'      => false,
 			)
 		);
 	}